Municipalities Brazil data analyis
This is a spatial data analyis project , i did in R in which iproduced a map showing the centroids of each municipality in just the state of Säo  Paulo, and add the outer boundary of Säo Paulo state. I also determined  the mean Human Development Index of Municipalities in each state of Brazil, using HDI as the variable from Municipalities data set.went on to produce a polygon/shapefile mapping the area of the municipality ‘Gaucha do Norte’ that is in the indigenous territory “Parque do Xingu”. Then In the state of Acre (AC), which two social housing (MCMV) projects are closest to each other, i had to create a 10km buffer around each housing project.